Separate the packing from the arrowsAlongsideChangeOutputs style.
