+++ lib/assertions.js
@@ -1044,7 +1044,8 @@ module.exports = function (expect) {
                             if (subjectType.indent) {
                                 output.indentLines();
                             }
-                            output.arrowsAlongsideChangeOutputs(changes, changes.map(function (diffItem, index) {
+                            var packing = utils.packArrows(changes);
+                            output.arrowsAlongsideChangeOutputs(packing, changes.map(function (diffItem, index) {
                                 var delimiterOutput = subjectType.delimiter(output.clone(), index, indexOfLastNonInsert + 1);
                                 return output.clone().block(function () {
                                     var type = diffItem.type;
