+++ lib/styles.js
@@ -1,7 +1,6 @@
 var utils = require('./utils');
 var stringDiff = require('diff');
 var specialCharRegExp = require('./specialCharRegExp');
-var greedyIntervalPacker = require('greedy-interval-packer');
 
 module.exports = function (expect) {
     expect.installTheme({ styles: {
@@ -419,45 +418,18 @@ module.exports = function (expect) {
         }
     });
 
-    expect.addStyle('arrowsAlongsideChangeOutputs', function (changes, changeOutputs) {
-        var moveSourceAndTargetByActualIndex = {};
-        changes.forEach(function (diffItem, index) {
-            if (diffItem.type === 'moveSource') {
-                diffItem.changeIndex = index;
-                (moveSourceAndTargetByActualIndex[diffItem.actualIndex] = moveSourceAndTargetByActualIndex[diffItem.actualIndex] || {}).source = diffItem;
-            } else if (diffItem.type === 'moveTarget') {
-                diffItem.changeIndex = index;
-                (moveSourceAndTargetByActualIndex[diffItem.actualIndex] = moveSourceAndTargetByActualIndex[diffItem.actualIndex] || {}).target = diffItem;
-            }
-        });
-
-        var topByChangeNumber = {};
-        var top = 0;
-        changeOutputs.forEach(function (changeOutput, index) {
-            topByChangeNumber[index] = top;
-            top += changeOutput.size().height;
-        });
-        var that = this;
-        var moveIndices = Object.keys(moveSourceAndTargetByActualIndex);
-        if (moveIndices.length > 0) {
-            var arrowSpecs = [];
-            moveIndices.sort(function (a, b) {
-                // Order by distance between change indices descending
-                return Math.abs(moveSourceAndTargetByActualIndex[b].source.changeIndex - moveSourceAndTargetByActualIndex[b].target.changeIndex) - Math.abs(moveSourceAndTargetByActualIndex[a].source.changeIndex - moveSourceAndTargetByActualIndex[a].target.changeIndex);
-            }).forEach(function (actualIndex, i, keys) {
-                var moveSourceAndMoveTarget = moveSourceAndTargetByActualIndex[actualIndex];
-                var firstChangeIndex = Math.min(moveSourceAndMoveTarget.source.changeIndex, moveSourceAndMoveTarget.target.changeIndex);
-                var lastChangeIndex = Math.max(moveSourceAndMoveTarget.source.changeIndex, moveSourceAndMoveTarget.target.changeIndex);
-
-                arrowSpecs.push({
-                    start: firstChangeIndex,
-                    end: lastChangeIndex,
-                    direction: moveSourceAndMoveTarget.source.changeIndex < moveSourceAndMoveTarget.target.changeIndex ? 'down' : 'up'
-                });
+    expect.addStyle('arrowsAlongsideChangeOutputs', function (packing, changeOutputs) {
+        if (packing) {
+            var topByChangeNumber = {};
+            var top = 0;
+            changeOutputs.forEach(function (changeOutput, index) {
+                topByChangeNumber[index] = top;
+                top += changeOutput.size().height;
             });
+            var that = this;
 
             var arrows = [];
-            greedyIntervalPacker(arrowSpecs).forEach(function (columnSet, i, packing) {
+            packing.forEach(function (columnSet, i, packing) {
                 columnSet.forEach(function (entry) {
                     arrows.push(that.clone().arrow({
                         left: i * 2,
