+++ lib/types.js
@@ -557,7 +557,7 @@ module.exports = function (expect) {
             var indexOfLastNonInsert = changes.reduce(function (previousValue, diffItem, index) {
                 return (diffItem.type === 'insert') ? previousValue : index;
             }, -1);
-            output.arrowsAlongsideChangeOutputs(changes, changes.map(function (diffItem, index) {
+            output.arrowsAlongsideChangeOutputs(utils.packArrows(changes), changes.map(function (diffItem, index) {
                 var delimiterOutput = type.delimiter(output.clone(), index, indexOfLastNonInsert + 1);
                 return output.clone().block(function () {
                     var type = diffItem.type;
