+++ lib/utils.js
@@ -1,5 +1,6 @@
 /* eslint-disable no-proto */
 var canSetPrototype = Object.setPrototypeOf || ({ __proto__: [] } instanceof Array);
+var greedyIntervalPacker = require('greedy-interval-packer');
 
 var setPrototypeOf = Object.setPrototypeOf || function setPrototypeOf(obj, proto) {
     obj.__proto__ = proto;
@@ -147,5 +148,38 @@ var utils = module.exports = {
         }, Array.prototype.slice.call(arguments));
     },
 
-    numericalRegExp: /^(?:0|[1-9][0-9]*)$/
+    numericalRegExp: /^(?:0|[1-9][0-9]*)$/,
+
+    packArrows: function (changes) {
+        var moveSourceAndTargetByActualIndex = {};
+        changes.forEach(function (diffItem, index) {
+            if (diffItem.type === 'moveSource') {
+                diffItem.changeIndex = index;
+                (moveSourceAndTargetByActualIndex[diffItem.actualIndex] = moveSourceAndTargetByActualIndex[diffItem.actualIndex] || {}).source = diffItem;
+            } else if (diffItem.type === 'moveTarget') {
+                diffItem.changeIndex = index;
+                (moveSourceAndTargetByActualIndex[diffItem.actualIndex] = moveSourceAndTargetByActualIndex[diffItem.actualIndex] || {}).target = diffItem;
+            }
+        });
+        var moveIndices = Object.keys(moveSourceAndTargetByActualIndex);
+        if (moveIndices.length > 0) {
+            var arrowSpecs = [];
+            moveIndices.sort(function (a, b) {
+                // Order by distance between change indices descending
+                return Math.abs(moveSourceAndTargetByActualIndex[b].source.changeIndex - moveSourceAndTargetByActualIndex[b].target.changeIndex) - Math.abs(moveSourceAndTargetByActualIndex[a].source.changeIndex - moveSourceAndTargetByActualIndex[a].target.changeIndex);
+            }).forEach(function (actualIndex, i, keys) {
+                var moveSourceAndMoveTarget = moveSourceAndTargetByActualIndex[actualIndex];
+                var firstChangeIndex = Math.min(moveSourceAndMoveTarget.source.changeIndex, moveSourceAndMoveTarget.target.changeIndex);
+                var lastChangeIndex = Math.max(moveSourceAndMoveTarget.source.changeIndex, moveSourceAndMoveTarget.target.changeIndex);
+
+                arrowSpecs.push({
+                    start: firstChangeIndex,
+                    end: lastChangeIndex,
+                    direction: moveSourceAndMoveTarget.source.changeIndex < moveSourceAndMoveTarget.target.changeIndex ? 'down' : 'up'
+                });
+            });
+
+            return greedyIntervalPacker(arrowSpecs);
+        }
+    }
 };
